Changed the update interval from 10ms to 30ms.
The behavior of some demos (Blob Sallad, Starfield etc.) became unstable as the rendering took longer time. I don't know an exact reason, but this modification stabilize the behavior of those demos.

Note: The minimum resolution of the timer IE uses is 15.6ms. Namely, the real meaning of this modification is to change the interval from 15.6ms to 31.2ms.


git-svn-id: http://flashcanvas.googlecode.com/svn/trunk@9 0a2e34be-cbc4-11de-bd98-e341d93cc3cc