js: Fixed a bug that resize event canceled createLinearGradient().
When resize event is fired, the commands queued in the buffer are
cleared and not sent to Flash. This is fine in most cases since the
bitmap of Canvas is also initialized by resize event.

However, if an execution of createLinearGradient() is canceled, it
causes an inconsistency: CanvasGradient object is not created in Flash
side while it exists in JavaScript side.


git-svn-id: http://flashcanvas.googlecode.com/svn/trunk@87 0a2e34be-cbc4-11de-bd98-e341d93cc3cc